refactor: extract fetch wrapper and browser open logic
- extracted createAntigravityFetch to src/plugin/fetch-wrapper.ts
- extracted openBrowser to src/plugin/browser.ts
- added MAX_ACCOUNTS constant
- simplified src/plugin.ts
- updated executeSearch to use abort signal from context
