+++ src/constants.ts
@@ -49,3 +49,5 @@ export const SEARCH_MODEL = "gemini-2.5-flash";
 export const SEARCH_THINKING_BUDGET_FAST = 4096;
 export const SEARCH_THINKING_BUDGET_DEEP = 16384;
 export const SEARCH_TIMEOUT_MS = 60 * 1000;
+
+export const MAX_ACCOUNTS = 10;
